Use embedded rule testing for MetaTagValidRule
Partial-Fix: #947
