+++ lib/ansiblelint/rules/MetaTagValidRule.py
@@ -1,6 +1,8 @@
+#!/usr/bin/env python
 # Copyright (c) 2018, Ansible Project
 
 import re
+import sys
 
 from ansiblelint.rules import AnsibleLintRule
 
@@ -57,3 +59,28 @@ class MetaTagValidRule(AnsibleLintRule):
                                 "{}, invalid: '{}'".format(msg, tag)))
 
         return results
+
+
+META_TAG_VALID = '''
+galaxy_info:
+    galaxy_tags: ['database', 'my s q l', 'MYTAG']
+    categories: 'my_category_not_in_a_list'
+'''
+
+# testing code to be loaded only with pytest or when executed the rule file
+if "pytest" in sys.modules or __name__ == "__main__":
+
+    import pytest
+
+    # Allows us to execute the file to run the tests insite it
+    if __name__ == "__main__":
+        pytest.main()
+
+    @pytest.mark.parametrize('rule_runner', (MetaTagValidRule, ), indirect=['rule_runner'])
+    def test_valid_tag_rule(rule_runner):
+        """Test rule matches."""
+        results = rule_runner.run_role_meta_main(META_TAG_VALID)
+        assert "Use 'galaxy_tags' rather than 'categories'" in str(results)
+        assert "Expected 'categories' to be a list" in str(results)
+        assert "invalid: 'my s q l'" in str(results)
+        assert "invalid: 'MYTAG'" in str(results)
