+++ pytest.ini
@@ -54,5 +54,8 @@ python_files =
     # Ref: https://docs.pytest.org/en/latest/reference.html#confval-python_files
     # Needed to discover legacy nose test modules:
     Test*.py
+    # Needed to discover embedded Rule tests
+    *Rule.py
+# ansiblelint/rules is added dynamically to the testpaths to include embedded tests
 testpaths = test/
 xfail_strict = true
